pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hormuz-transit-index"
version = "0.1.0"
description = "24/7 AIS-based Hormuz Strait transit monitoring and index generation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.0",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
